rename the builder macros to include a "fmp" prefix
As a preparation for adding S2I flow tests, the builder macros
for running the Fabric8 Maven plugin flow tests are renamed
to include a `fmp-` prefix. The content of those macros is now
externalized to separate shell scripts, which are easier to edit.

Also the JSON job descriptions are no longer parameterized
with the "type" field, that is now hardcoded. That actualy reduces
code duplication and it will also make it much easier to add jobs
for the S2I flow tests.